read, get_ott, nickname code complete with tests
