+++ tests/test_io.py
@@ -4,73 +4,7 @@ from topiary import io
 import numpy as np
 import pandas as pd
 
-import warnings
-
-def test_check_topiary_dataframe(test_dataframes):
-    """
-    Test check for topiary dataframe.
-    """
-
-    # Make sure reads good dataframe without mangling
-    good_df = test_dataframes["good-df"]
-    df = io.check_topiary_dataframe(test_dataframes["good-df"])
-    assert np.sum(np.asarray(good_df == df) == False) == 0
-
-    # make sure it properly deals with all sorts of wacky input
-    bad_inputs = [1,-1,1.5,None,False,pd.DataFrame]
-    for b in bad_inputs:
-        with pytest.raises(ValueError):
-            io.check_topiary_dataframe(b)
-
-    # Make sure it drops empty lines
-    with pytest.warns():
-        df = io.check_topiary_dataframe(test_dataframes["good-test_blank-lines"])
-    assert len(df) == 5
-
-    # Make sure it properly looks for required columns
-    required_col = ["species","name","sequence"]
-    for c in required_col:
-        bad_df = good_df.copy()
-        bad_df = bad_df.drop(columns=[c])
-        with pytest.raises(ValueError):
-            io.check_topiary_dataframe(bad_df)
-
-    # Check keep
-    with pytest.warns():
-        df = io.check_topiary_dataframe(test_dataframes["good-test-keep-parse"])
-    assert df.keep.dtypes is np.dtype(bool)
-    assert np.array_equal(df.keep,np.array([True,False,
-                                            True,False,
-                                            True,False]))
-
-    with pytest.warns():
-        df = io.check_topiary_dataframe(test_dataframes["good-test-keep-parse_number"])
-    assert df.keep.dtypes is np.dtype(bool)
-    assert np.array_equal(df.keep,np.array([True,False,
-                                            True,False]))
-
-
-    # Check add uid column. Should warn that it's editing uid
-    with pytest.warns():
-        df = io.check_topiary_dataframe(test_dataframes["no-uid"])
-    assert len(np.unique(df.uid)) == len(df)
-
-    # Check replace bad uid. Should warn that it's editing uid
-    with pytest.warns():
-        df = io.check_topiary_dataframe(test_dataframes["bad-uid"])
-    for d in df.uid:
-        assert type(d) is str
-        assert len(d) == 10
-
-    # Check make uid unique. Should warn that it's editing uid
-    with pytest.warns():
-        df = io.check_topiary_dataframe(test_dataframes["duplicate-uid"])
-    assert len(np.unique(df.uid)) == len(df)
-
-
-    # Check ott
-
-
+import warnings, os
 
 
 
@@ -113,6 +47,7 @@ def test_read_dataframe(dataframe_good_files,test_dataframes):
     df = io.read_dataframe("junk.csv",remove_extra_index=True)
     assert not df.columns[0].startswith("Unnamed:")
     assert len(df) == len(ref_df)
+    os.remove("junk.csv")
 
     # Make sure dies with useful error
     bad_inputs = [1,-1,1.5,None,False,pd.DataFrame]
@@ -123,6 +58,10 @@ def test_read_dataframe(dataframe_good_files,test_dataframes):
     # Make sure raises file not found if a file is not passed
     with pytest.raises(FileNotFoundError):
         io.read_dataframe("not_really_a_file.txt")
+
+def test_write_dataframe():
+    pass
+
