+++ topiary/io.py
@@ -18,262 +18,10 @@ from Bio import SeqIO
 
 from tqdm.auto import tqdm
 
-import re, string, random, io, warnings
+import re, string, random, io, warnings, copy
 
 
-def check_topiary_dataframe(df):
-    """
-    Check to make sure topiary dataframe is sane. Edits dataframe, returning a
-    copy.
 
-    + Checks for required columns 'species', 'name', and 'sequence'
-    + Deletes duplicated rows.
-    + Deletes empty rows.
-
-    + 'keep':
-        + Must be boolean.
-        + If present but not boolean, try to coerce it into boolean.
-            + if str, map regex [1yt] --> True and regex[0nf] --> False
-            + if int, map !0 --> True and 0 --> False
-            + if float, map !~0 --> True and ~0 --> False
-        + If keep is not present, add and set all to True.
-    + 'uid':
-        + Must be unique str of length 10 consisting only of [a-z][A-Z]
-        + If present:
-            + If missing values, fill them in
-            + If values do not meet rules, replace them with uid that do
-            + If values are not unique, update them to be unique
-        + If not present, create them.
-    + 'ott': If present, makes sure it has the format ottINTEGER.
-    """
-
-    # Make sure type is right
-    if type(df) is not pd.DataFrame:
-        err = "\ndf must be a pandas dataframe.\n\n"
-        raise ValueError(err)
-
-    # Make sure the dataframe has all required columns
-    required_columns = ["species","name","sequence"]
-    for c in required_columns:
-
-        try:
-            df[c]
-        except KeyError:
-            err = f"\n\nA topiary dataframe must have a {c} column.\n\n"
-            raise ValueError(err)
-
-    # Drop duplicates
-    df = df.drop_duplicates(ignore_index=True)
-
-    # Get mask for cells that are either null or nan
-    bad_mask = np.logical_or(pd.isnull(df),pd.isna(df))
-    idx = list(df.index)
-    final_mask = []
-    for i in range(len(bad_mask)):
-
-        # Grab row
-        row = df.loc[idx[i],:]
-
-        # Get all non-null and non-na entries from the row
-        mask = np.logical_not(np.logical_or(pd.isnull(row),pd.isna(row)))
-        not_null = row[mask]
-
-        # Default to not keeping a row
-        keep = False
-
-        # If there are any non-null entries left...
-        if len(not_null) > 0:
-
-            # Loop will end up with keep = False if only columns remaining are
-            # type string and "".
-            for n in not_null:
-
-                # If an empty string, this is basically a null.
-                if type(n) is str and n.strip() == "":
-                    continue
-
-                # If we get here, we have some kind of non-empty cell. Break out
-                # and keep row
-                keep = True
-                break
-
-        final_mask.append(keep)
-
-    final_mask = np.array(final_mask,dtype=bool)
-
-    # Let user know we're dropping lines
-    if np.sum(np.logical_not(final_mask)) > 0:
-        lines = ",".join(["{}".format(i) for i in df.index[np.logical_not(final_mask)]])
-        w = f"\nDropping apparently empty lines ({lines})\n\n"
-        warnings.warn(w)
-
-    # Drop rows that are all empty
-    df = df.loc[df.index[final_mask],:]
-
-    # -------------------------------------------------------------------------
-    # Process keep column
-
-    # Try to extract keep column
-    try:
-        keep = df.loc[:,"keep"]
-    # No column: create one and set all to True
-    except KeyError:
-        keep = None
-        df["keep"] = np.ones(len(df),dtype=bool)
-
-    if keep is not None:
-
-        # Do a pass trying to infer the datatype of keep. (This is useful if we
-        # dropped empty rows that made the original pandas read this column in
-        # as a mix of bool and object).
-        df.loc[:,"keep"] = df.keep.infer_objects()
-
-        # If it's not a boolean column, try to turn into one
-        if not np.dtype(keep.dtypes) is np.dtype(bool):
-
-            # Base message. If everything works great, let user know what
-            # happened as warning. If things go awry, use as start of error
-            # message
-            w = "\n\n"
-            w += "The 'keep' column must be boolean (True/False). pandas did\n"
-            w += "not recognize the column as boolean, so we're parsing it\n"
-            w += "manually by looking for 0/1, yes/no, true/false, etc.\n\n"
-
-            new_keep = []
-            look_for_true = re.compile("[1yt]",re.IGNORECASE)
-            look_for_false = re.compile("[0nf]",re.IGNORECASE)
-            for k in df.keep:
-                if type(k) is bool:
-                    is_true = True and k
-                    is_false = not is_true
-                    looks_like_a = "bool"
-                elif type(k) is str:
-                    is_true = look_for_true.search(k) is not None
-                    is_false = look_for_false.search(k) is not None
-                    looks_like_a = "string"
-                elif type(k) is int:
-                    is_true = (k != 0)
-                    is_false = (k == 0)
-                    looks_like_a = "int"
-                elif type(k) is float:
-                    is_true = np.logical_not(np.isclose(k,0))
-                    is_false = np.isclose(k,0)
-                    looks_like_a = "float"
-                else:
-                    w += f"Could not figure out how to parse '{k}'\n\n"
-                    raise ValueError(w)
-
-                if (is_true and is_false) or (not is_true and not is_false):
-                    w += f"Trying to parse '{k}' as a {looks_like_a}, but\n"
-                    w += f"could not figure out whether true or false.\n\n"
-                    raise ValueError(w)
-                else:
-                    new_keep.append(is_true)
-
-            # Record newly boolean-ized values
-            df.loc[:,"keep"] = np.array(new_keep,dtype=bool)
-
-            # Let user know we manually parsed the keep column...
-            warnings.warn(w)
-
-    # -------------------------------------------------------------------------
-    # Process uid column
-
-    warn_uid = ""
-
-    # Make sure the dataframe has a uid column
-    try:
-        uid = df.loc[:,"uid"]
-    except KeyError:
-        uid =  None
-        warn_uid += "  + No 'uid' column found. Creating column.\n"
-        df["uid"] = _private.generate_uid(len(df))
-
-    if uid is not None:
-
-        # Look for a-z and A-Z (only allowed characters in uid)
-        p = re.compile("[^a-z]",re.IGNORECASE)
-
-        final_uid = []
-
-        # Go through uid...
-        for u in uid:
-
-            # Force to be a spring
-            if type(u) is not str:
-                u = str(u)
-
-            # Disallowed character(s) in uid or uid not right length
-            if p.search(u) or len(u) != 10:
-
-                new_uid = _private.generate_uid()
-                warn_uid += f"  + uid '{u}' is invalid. Replacing with '{new_uid}'\n"
-                final_uid.append(new_uid)
-
-            else:
-                final_uid.append(u)
-
-        df.loc[:,"uid"] = final_uid
-
-    # Make sure uid are unique
-    uid, counts = np.unique(df.loc[:,"uid"],return_counts=True)
-    for u in uid[counts != 1]:
-        warn_uid = f"  + uid '{u}' is not unique. This will be replaced with a\n"
-        warn_uid += "    new unique uid for each sequence.\n"
-
-        mask = df.loc[:,"uid"] == u
-        df.loc[mask,"uid"] = _private.generate_uid(sum(mask))
-
-    if warn_uid != "":
-        w = "\n\n"
-        w += "'uid' column was invalid. topiary has fixed the problems noted below.\n"
-        w += "If you have already generated phylogenetic trees using a previous\n"
-        w += "version of this dataframe, **those trees are no longer compatible\n"
-        w += "with this dataframe**. To ensure compatibility, fix the problems\n"
-        w += "noted below and then re-read the dataframe into topiary. If you\n"
-        w += "have not already generated trees (or plan to generate new ones from\n"
-        w += "scratch) you may safely disregard this warning.\n"
-        w += "\n"
-        w += warn_uid
-        warnings.warn(w)
-
-    # -------------------------------------------------------------------------
-    # Check format of ott column if present
-
-    try:
-        ott = df.loc[:,"ott"]
-    except KeyError:
-        ott = None
-
-    if ott is not None:
-
-        for index in df.index:
-
-            # Get ott id
-            o = df.loc[index,"ott"]
-
-            # Missing value is okay --> make sure it's a pandas NULL datatype
-            if pd.isna(o) or pd.isnull(o) or o == "":
-                df.loc[index,"ott"] = pd.NA
-                continue
-
-            # If there is an ott that is not a null, make sure it is sane and
-            # reasable.
-            failed = False
-            if type(o) is str and o[:3] == "ott":
-                try:
-                    int(o[3:])
-                except ValueError:
-                    failed = True
-            else:
-                failed = True
-
-            if failed:
-                err = "\n\nott column must have format 'ottINTEGER', where \n"
-                err += "INTEGER is a the integer OTT accession number.\n"
-                raise ValueError(err)
-
-    return df
 
 def read_dataframe(input,remove_extra_index=True):
     """
@@ -324,7 +72,7 @@ def read_dataframe(input,remove_extra_index=True):
                     df = df.drop(columns=[df.columns[0]])
 
     # Validate the dataframe
-    df = check_topiary_dataframe(df)
+    df = util.check_topiary_dataframe(df)
 
     return df
 
@@ -357,9 +105,8 @@ def write_dataframe(df,out_file):
         df.to_excel(out_file,index=False)
 
 
-def ncbi_blast_xml_to_df(xml_files,
-                         aliases=None,
-                         phylo_context="All life"):
+
+def ncbi_blast_xml_to_df(xml_files):
     """
     Take a list of blast xml files and load in all sequences as a single
     topiary data frame. Parse meta data in an intelligent way, download
@@ -368,38 +115,6 @@ def ncbi_blast_xml_to_df(xml_files,
 
     xml_files: blast xml files to load. if a string, treat as a single xml file.
                if a list, treat as a list of xml files.
-    aliases: dictionary for standardizing protein names.  Key specifies what
-             should be output, values degenerate names that map back to that
-             key.  For example:
-                 "S100A9":("S100-A9","S100 A9","S-100 A9")
-             would replace "S100-A9", "S100 A9", and "S-100 A9" with "S100A9"
-    phylo_context: string. used to limit species seach for looking up species
-                   ids on open tree of life.  To get latest strings recognized
-                   by the database, use the following code:
-
-                   ```
-                   from opentree import OT
-                   print(OT.tnrs_contexts().response_dict)
-                   ```
-
-                   As of 2021-08-16, the following are recognized. You can use
-                   either the keys or values in this dictionary.
-
-                   {'ANIMALS': ['Animals','Birds','Tetrapods','Mammals',
-                                'Amphibians','Vertebrates','Arthropods',
-                                'Molluscs','Nematodes','Platyhelminthes',
-                                'Annelids','Cnidarians','Arachnids','Insects'],
-                    'FUNGI': ['Fungi', 'Basidiomycetes', 'Ascomycetes'],
-                    'LIFE': ['All life'],
-                    'MICROBES': ['Bacteria','SAR group','Archaea','Excavata',
-                                 'Amoebozoa','Centrohelida','Haptophyta',
-                                 'Apusozoa','Diatoms','Ciliates','Forams'],
-                    'PLANTS': ['Land plants','Hornworts','Mosses','Liverworts',
-                               'Vascular plants','Club mosses','Ferns',
-                               'Seed plants','Flowering plants','Monocots',
-                               'Eudicots','Rosids','Asterids','Asterales',
-                               'Asteraceae','Aster','Symphyotrichum',
-                               'Campanulaceae','Lobelia']}
 
     returns a pandas data frame
     """
@@ -429,8 +144,7 @@ def ncbi_blast_xml_to_df(xml_files,
             start = tmp_df.loc[j,"subject_start"]
             end = tmp_df.loc[j,"subject_end"]
             hit_info = ncbi.parse_ncbi_line(title,
-                                            accession=accession,
-                                            aliases=aliases)
+                                            accession=accession)
             if hit_info is None:
                 continue
 
@@ -465,6 +179,13 @@ def ncbi_blast_xml_to_df(xml_files,
     # Create a dictionary with appropriate keys to load into dataframe.
     out = util.create_pipeline_dict()
 
+    # Add blast-specific columns
+    blast_key_list = ["accession","xml","length","evalue","start","end",
+                      "structure","low_quality","precursor","predicted",
+                      "isoform","hypothetical","partial","raw_line"]
+    for k in blast_key_list:
+        out[k] = []
+
     # Go through every hit
     for i in range(len(all_hits)):
 
@@ -496,13 +217,12 @@ def ncbi_blast_xml_to_df(xml_files,
         out["start"].append(start)
         out["end"].append(end)
 
-        out["uid"].append("".join([random.choice(string.ascii_letters) for _ in range(10)]))
+        out["uid"].append(_private.generate_uid())
 
         out["keep"].append(True)
 
     df = pd.DataFrame(out)
-
-    df = opentree.get_ott_id(df,context_name=phylo_context)
+    df = util.check_topiary_dataframe(df)
 
     return df
